+++ pkg/crd/apply_test.go
@@ -55,6 +55,28 @@ func dropLogMessages(t *testing.T, s tests.LogScanner) map[string]string {
 }
 
 func Test_Apply(t *testing.T) {
+	t.Run("NoSchema", func(t *testing.T) {
+		crdValidation := make(map[string]crds.CRDOptions)
+		for n := range GetDefaultCRDOptions() {
+			crdValidation[n] = crds.CRDOptions{
+				WithSchema: false,
+			}
+		}
+		runApply(t, crdValidation)
+	})
+	t.Run("WithSchema", func(t *testing.T) {
+		crdValidation := make(map[string]crds.CRDOptions)
+		for n := range GetDefaultCRDOptions() {
+			crdValidation[n] = crds.CRDOptions{
+				WithSchema: true,
+			}
+		}
+		runApply(t, crdValidation)
+	})
+}
+
+func runApply(t *testing.T, crdOpts map[string]crds.CRDOptions) {
+	t.Helper()
 	verifyCRDAccessForTests = &authorization.SubjectAccessReviewStatus{
 		Allowed: true,
 	}
@@ -66,7 +88,7 @@ func Test_Apply(t *testing.T) {
 			c := kclient.NewFakeClient()
 
 			t.Run("Ensure", func(t *testing.T) {
-				require.NoError(t, EnsureCRD(context.Background(), c, false))
+				require.NoError(t, EnsureCRDWithOptions(context.Background(), c, EnsureCRDOptions{IgnoreErrors: false, CRDOptions: crdOpts}))
 
 				for k, v := range dropLogMessages(t, s) {
 					t.Run(k, func(t *testing.T) {
@@ -92,7 +114,7 @@ func Test_Apply(t *testing.T) {
 			})
 
 			t.Run("Ensure", func(t *testing.T) {
-				require.NoError(t, EnsureCRD(context.Background(), c, false))
+				require.NoError(t, EnsureCRDWithOptions(context.Background(), c, EnsureCRDOptions{IgnoreErrors: false, CRDOptions: crdOpts}))
 
 				for k, v := range dropLogMessages(t, s) {
 					t.Run(k, func(t *testing.T) {
@@ -120,7 +142,7 @@ func Test_Apply(t *testing.T) {
 			})
 
 			t.Run("Ensure", func(t *testing.T) {
-				require.NoError(t, EnsureCRD(context.Background(), c, false))
+				require.NoError(t, EnsureCRDWithOptions(context.Background(), c, EnsureCRDOptions{IgnoreErrors: false, CRDOptions: crdOpts}))
 
 				for k, v := range dropLogMessages(t, s) {
 					t.Run(k, func(t *testing.T) {